refactor(frontend): integrate MudSocketAdapter for enhanced WebSocket handling and message transformation
