+++ frontend/src/app/features/terminal/index.ts
@@ -1,3 +1,4 @@
 export * from './models/escapes';
 export * from './mud-input.controller';
 export * from './mud-prompt.manager';
+export * from './mud-socket.adapter';
